requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client>=5.0.0
pypdf
python-dotenv
requests
langchain-text-splitters